# internal/docs/swagger.yaml
basePath: /api
definitions:
  httpsuite.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  models.AlarmRule:
    properties:
      enabled:
        type: boolean
      id:
        type: string
      last_fired:
        type: string
      last_value:
        type: number
      message:
        type: string
      name:
        type: string
      queue_name:
        type: string
      status:
        $ref: '#/definitions/models.AlarmStatus'
      threshold:
        type: number
      type:
        $ref: '#/definitions/models.AlarmType'
    type: object
  models.AlarmRuleCreate:
    properties:
      enabled:
        example: true
        type: boolean
      message:
        example: Queue size has exceeded the threshold
        type: string
      name:
        example: High Queue Size
        type: string
      queue_name:
        example: my-queue
        type: string
      threshold:
        example: 1000
        type: number
      type:
        allOf:
        - $ref: '#/definitions/models.AlarmType'
        example: queue_size
    type: object
  models.AlarmRulePatch:
    properties:
      enabled:
        example: true
        type: boolean
      message:
        example: Queue size has exceeded the threshold
        type: string
      name:
        example: High Queue Size
        type: string
      queue_name:
        $ref: '#/definitions/httpsuite.Optional-string'
      threshold:
        example: 1000
        type: number
      type:
        example: queue_size
        type: string
    type: object
  models.AlarmStatus:
    enum:
    - ok
    - active
    - inactive
    - firing
    - fired
    - unknown
    type: string
    x-enum-varnames:
    - OK
    - Active
    - Inactive
    - Firing
    - Fired
    - AlarmStatusUnknown
  models.AlarmType:
    enum:
    - channels
    - connections
    - queues
    - unacked
    - queue_messages
    - queue_size
    - no_consumer
    - maintenance
    type: string
    x-enum-varnames:
    - channels
    - connections
    - queues
    - unacked_messages
    - queue_messages
    - queue_size
    - no_consumer
    - maintenance
...
  models.TestNotificationResponse:
    properties:
      failed_destinations:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: Norsk helsenett SF
    url: https://github.com/NorskHelsenett/unimq-dashboard
  description: API for RabbitMQ Dashboard application.
  title: RabbitMQ Dashboard API
  version: "1.0"
paths:
...
  /v1/notifications/{vhost-name}/rules:
    post:
      consumes:
      - application/json
      description: Add a new notification rule for a specific vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Notification Rule Object
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/models.AlarmRuleCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Rule added successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Add a new notification rule
      tags:
      - Notifications
  /v1/notifications/{vhost-name}/rules/{rule-id}:
    delete:
      description: Delete a specific notification rule for a vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Notification Rule ID
        in: path
        name: rule-id
        required: true
        type: string
      responses:
        "200":
          description: Rule deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Delete a notification rule
      tags:
      - Notifications
    get:
      description: Retrieve a specific notification rule for a vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Notification Rule ID
        in: path
        name: rule-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Notification rule retrieved successfully
          schema:
            $ref: '#/definitions/models.AlarmRule'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Get a notification rule
      tags:
      - Notifications
    patch:
      description: Patch a specific notification rule for a vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Notification Rule ID
        in: path
        name: rule-id
        required: true
        type: string
      - description: Patched Notification Rule Object
        in: body
        name: rule
        required: true
        schema:
          $ref: '#/definitions/models.AlarmRulePatch'
      responses:
        "200":
          description: Rule patched successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Patch a notification rule
      tags:
      - Notifications
  /v1/notifications/{vhost-name}/rules/{rule-id}/test:
    post:
      description: Send a test notification using the specified rule to verify its
        configuration
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Notification Rule ID
        in: path
        name: rule-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Test notification sent successfully
          schema:
            $ref: '#/definitions/models.TestNotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Test Notification Rule
      tags:
      - Notifications
  /v1/notifications/{vhost-name}/rules/{rule-id}/toggle:
    post:
      description: Enable or disable a specific notification rule for a vhost
      parameters:
      - description: Vhost Name
        in: path
        name: vhost-name
        required: true
        type: string
      - description: Notification Rule ID
        in: path
        name: rule-id
        required: true
        type: string
      responses:
        "200":
          description: Rule toggled successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpsuite.ErrorResponse'
      security:
      - bearer: []
      - OAuth2:
        - openid
        - profile
        - email
        - groups
        - audience:server:client_id:unimq-dashboard
      summary: Toggle a notification rule
      tags:
      - Notifications
...
securityDefinitions:
  OAuth2:
    authorizationUrl: http://localhost:5556/dex/auth
    flow: accessCode
    scopes:
      audience:server:client_id:unimq-dashboard: Issue a token whose audience the
        dashboard API accepts
      email: User email address
      groups: Group membership used for authorization
      openid: OpenID Connect identity
      profile: User profile
    tokenUrl: http://localhost:5556/dex/token
    type: oauth2
  bearer:
    description: '"JWT token for authentication, obtained from the Dex OIDC provider,
      formatted as ''Bearer {token}'' in the Authorization header"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"